Add optional osmp-binary-variable annotations
This change introduces optional osmp-binary-variable annotations in
order to support generic binary variables in the OSMP format.  This
extension should allow simpler support for e.g. raw data streams to be
added later on and should ease porting to actual binary variables if
they are added to the FMI standard proper.