Add flag to allow linking against shared OSI lib
This currently only works on Linux, since the default OSI build does
not generate the needed export library on Windows, and it is currently
also only needed for certain Linux usages, since multiple static
instances of open_simulation_interface tend to clash on Linux, but not
on Windows.  The flag defaults to OFF for now (default might change on
Linux in the future).