Fix source FMU generation via conditional prefix
The generation of the source FMU was flawed since the needed prefix
macro FMI2_FUNCTION_PREFIX was not being added to the distributed
sources. Since this is difficult to achieve portably with cmake, we
switch to including the prefix in the header always, and selective
deactivating it through compiler defines during build time for shared
onject/DLL builds.